feat: sep. config command
