+++ packages/cli/src/commands/install/install.ts
@@ -8,8 +8,8 @@
 import {logger} from '@react-native-community/cli-tools';
 import * as PackageManager from '../../tools/packageManager';
 import link from '../link/link';
-import loadConfig from '@react-native-community/cli-config';
 import {Config} from '@react-native-community/cli-types';
+import loadConfig from '@react-native-community/cli-config';
 
 async function install(args: Array<string>, ctx: Config): Promise<void> {
   const name = args[0];
